New server architecture - part 1
