+++ package.json
@@ -1,13 +1,12 @@
 {
-  "dependencies": {
-    "bcryptjs": "^2.4.3",
-    "config": "^3.3.6",
-    "express": "^4.17.1",
-    "express-validator": "^6.12.0",
-    "jsonwebtoken": "^8.5.1",
-    "mongoose": "^5.12.13"
-  },
+  "name": "the-psalmist",
+  "version": "1.0.0",
   "devDependencies": {
     "concurrently": "^6.2.0"
+  },
+  "scripts": {
+    "start-server": "cd the-psalmist-server && npm run server",
+    "start-client": "cd the-psalmist-client && npm run dev",
+    "start": "concurrently \"npm run start-server \" \"npm run start-client\""
   }
 }
