+++ the-psalmist-client/public/manifest.json
