+++ the-psalmist-client/src/components/layout/PreLoader.js
