+++ the-psalmist-server/utils/errors.js
@@ -1,0 +1,5 @@
+const makeError = (statusCode, msg) => ({
+    statusCode, msg
+})
+
+module.exports = makeError
